+++ src/test/resources/config.properties
@@ -1,3 +1,3 @@
 browser=chrome
 headless=false
-timeout=1
+url=https://www.saucedemo.com/
